Conflicts:
	entity-management-common/pom.xml
	entity-management-web/src/main/java/eu/europeana/entitymanagement/batch/processor/EntityConsolidationProcessor.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/batch/reader/EntityRecordDatabaseReader.java
	entity-management-zoho/src/main/java/eu/europeana/entitymanagement/wikidata/WikidataDereferenceService.java

+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/web/service/EntityRecordService.java
@@ -10,12 +10,11 @@ import static java.time.Instant.now;
 import com.mongodb.client.result.UpdateResult;
 import dev.morphia.query.experimental.filters.Filter;
 import eu.europeana.api.commons.error.EuropeanaApiException;
-import eu.europeana.entitymanagement.batch.utils.BatchUtils;
 import eu.europeana.entitymanagement.common.config.DataSource;
 import eu.europeana.entitymanagement.common.config.EntityManagementConfiguration;
 import eu.europeana.entitymanagement.config.AppConfig;
 import eu.europeana.entitymanagement.config.DataSources;
-import eu.europeana.entitymanagement.definitions.exceptions.EntityCreationException;
+import eu.europeana.entitymanagement.definitions.exceptions.EntityModelCreationException;
 import eu.europeana.entitymanagement.definitions.model.Address;
 import eu.europeana.entitymanagement.definitions.model.Agent;
 import eu.europeana.entitymanagement.definitions.model.Aggregation;
@@ -28,6 +27,7 @@ import eu.europeana.entitymanagement.definitions.model.TimeSpan;
 import eu.europeana.entitymanagement.definitions.model.WebResource;
 import eu.europeana.entitymanagement.definitions.web.EntityIdDisabledStatus;
 import eu.europeana.entitymanagement.exception.EntityAlreadyExistsException;
+import eu.europeana.entitymanagement.exception.EntityCreationException;
 import eu.europeana.entitymanagement.exception.EntityNotFoundException;
 import eu.europeana.entitymanagement.exception.EntityRemovedException;
 import eu.europeana.entitymanagement.exception.HttpBadRequestException;
@@ -151,22 +151,26 @@ public class EntityRecordService {
     return entityRecordRepository.save(er);
   }
 
-  public List<? extends EntityRecord> saveBulkEntityRecords(List<? extends EntityRecord> records) {
-    return entityRecordRepository.saveBulk((List<EntityRecord>) records);
+  public List<EntityRecord> saveBulkEntityRecords(List<EntityRecord> records) {
+    return entityRecordRepository.saveBulk(records);
   }
 
   /**
-   * This method deletes entities permanently from database and from solr index if the deleteFromSolr flag is set to true
+   * This method deletes entities permanently from database and from solr index if the
+   * deleteFromSolr flag is set to true
+   *
    * @param entityIds the ids of the entities to be deleted
-   * @param deleteFromSolr flag indicating if the 
+   * @param deleteFromSolr flag indicating if the
    * @return the number of entities deleted into the database
    * @throws SolrServiceException if an error occurs when deleteing from solr
    */
-  public long deleteBulk(List<String> entityIds, boolean deleteFromSolr) throws SolrServiceException {
-    if(deleteFromSolr) {
+  public long deleteBulk(List<String> entityIds, boolean deleteFromSolr)
+      throws SolrServiceException {
+    if (deleteFromSolr && !entityIds.isEmpty()) {
       solrService.deleteById(entityIds, true);
+      logger.info("Deleted {} entityRecords from Solr: entityIds={}", entityIds.size(), entityIds);
     }
-    
+
     long deleteCount = entityRecordRepository.deleteBulk(entityIds);
     if (deleteCount > 0) {
       logger.info("Deleted {} entityRecords from database: entityIds={}", deleteCount, entityIds);
@@ -175,14 +179,13 @@ public class EntityRecordService {
   }
 
   /**
-   * This method sets the deleted field in the database, it does not remove from solr. This method needs to be used only within the batch item writer
-   * @deprecated this method does not remove from solr. When using this method, it must be ensured that the delete from solr is also invoked   
-   * @param entityRecords
+   * This method sets the deleted field in the database, it does not remove from solr. This method
+   * needs to be used only within the batch item writer
+   *
+   * @param entityIds list of records to disable
    */
-  @Deprecated
-  public void disableBulk(List<? extends EntityRecord> entityRecords) {
-    String[] entityIds = BatchUtils.getEntityIds(entityRecords);
-    UpdateResult updateResult = entityRecordRepository.disableBulk(List.of(entityIds));
+  public void disableBulk(List<String> entityIds) {
+    UpdateResult updateResult = entityRecordRepository.disableBulk(entityIds);
     logger.info("Deprecated {} entities: entityIds={}", updateResult.getModifiedCount(), entityIds);
   }
 
@@ -190,10 +193,12 @@ public class EntityRecordService {
    * Mark entity record as disabled and remove from solr
    *
    * @param er the entity record
-   * @param forceSolrCommit indicating if a solr commit should be explicitly (synchronuously) executed 
+   * @param forceSolrCommit indicating if a solr commit should be explicitly (synchronuously)
+   *     executed
    * @throws EntityUpdateException incase of solr access errors
    */
-  public void disableEntityRecord(EntityRecord er, boolean forceSolrCommit) throws EntityUpdateException {
+  public void disableEntityRecord(EntityRecord er, boolean forceSolrCommit)
+      throws EntityUpdateException {
     try {
       solrService.deleteById(List.of(er.getEntityId()), forceSolrCommit);
     } catch (SolrServiceException e) {
@@ -202,7 +207,6 @@ public class EntityRecordService {
     er.setDisabled(new Date());
     saveEntityRecord(er);
   }
-
 
   /**
    * Re-Enable an already existing entity record.
@@ -244,7 +248,7 @@ public class EntityRecordService {
   public long delete(String entityId) throws SolrServiceException {
     // delete from Solr before Mongo, so Solr errors won't leave DB in an inconsistent state
     solrService.deleteById(List.of(entityId), true);
-    
+
     return entityRecordRepository.deleteForGood(entityId);
   }
 
@@ -262,8 +266,8 @@ public class EntityRecordService {
    */
   public EntityRecord createEntityFromMigrationRequest(
       Entity europeanaProxyEntity, String type, String identifier)
-      throws EntityCreationException, EntityAlreadyExistsException, HttpBadRequestException,
-          HttpUnprocessableException {
+      throws EntityAlreadyExistsException, HttpBadRequestException, HttpUnprocessableException,
+          EntityModelCreationException {
 
     String externalProxyId = europeanaProxyEntity.getEntityId();
 
@@ -317,7 +321,12 @@ public class EntityRecordService {
     String externalProxyId = europeanaProxyEntity.getEntityId();
 
     Date timestamp = new Date();
-    Entity entity = EntityObjectFactory.createConsolidatedEntityObject(europeanaProxyEntity);
+    Entity entity;
+    try {
+      entity = EntityObjectFactory.createConsolidatedEntityObject(europeanaProxyEntity);
+    } catch (EntityModelCreationException e) {
+      throw new EntityCreationException(e.getMessage(), e);
+    }
 
     boolean isZohoOrg = ZohoUtils.isZohoOrganization(externalProxyId, datasourceResponse.getType());
     String entityId = generateEntityId(datasourceResponse, isZohoOrg);
@@ -373,29 +382,33 @@ public class EntityRecordService {
   public EntityProxy appendWikidataProxy(
       EntityRecord entityRecord, String wikidataProxyId, String entityType, Date timestamp)
       throws EntityCreationException {
-    // entity metadata will be populated during update task
-    Entity wikidataProxyEntity = EntityObjectFactory.createProxyEntityObject(entityType);
+    try {
+      // entity metadata will be populated during update task
+      Entity wikidataProxyEntity = EntityObjectFactory.createProxyEntityObject(entityType);
 
-    Optional<DataSource> wikidataDatasource = getDataSource(wikidataProxyId);
-    // exception is thrown in factory method if wikidataDatasource is empty
-    int proxyNr = entityRecord.getProxies().size();
-    EntityProxy wikidataProxy =
-        setExternalProxy(
-            wikidataProxyEntity,
-            wikidataProxyId,
-            entityRecord.getEntityId(),
-            wikidataDatasource.get(),
-            entityRecord,
-            timestamp,
-            proxyNr);
+      Optional<DataSource> wikidataDatasource = getDataSource(wikidataProxyId);
+      // exception is thrown in factory method if wikidataDatasource is empty
+      int proxyNr = entityRecord.getProxies().size();
+      EntityProxy wikidataProxy =
+          setExternalProxy(
+              wikidataProxyEntity,
+              wikidataProxyId,
+              entityRecord.getEntityId(),
+              wikidataDatasource.get(),
+              entityRecord,
+              timestamp,
+              proxyNr);
 
-    // add wikidata uri to entity sameAs
-    entityRecord.getEntity().addSameReferenceLink(wikidataProxyId);
-    // add to entityIsAggregatedBy
-    Aggregation isAggregatedBy = entityRecord.getEntity().getIsAggregatedBy();
-    updateEntityAggregatesList(isAggregatedBy, entityRecord, entityRecord.getEntityId());
+      // add wikidata uri to entity sameAs
+      entityRecord.getEntity().addSameReferenceLink(wikidataProxyId);
+      // add to entityIsAggregatedBy
+      Aggregation isAggregatedBy = entityRecord.getEntity().getIsAggregatedBy();
+      updateEntityAggregatesList(isAggregatedBy, entityRecord, entityRecord.getEntityId());
 
-    return wikidataProxy;
+      return wikidataProxy;
+    } catch (EntityModelCreationException e) {
+      throw new EntityCreationException(e.getMessage(), e);
+    }
   }
 
   String generateEntityId(Entity datasourceResponse, boolean isZohoOrg) {
@@ -623,7 +636,8 @@ public class EntityRecordService {
    *
    * @throws EntityCreationException
    */
-  public Entity mergeEntities(Entity primary, Entity secondary) throws EuropeanaApiException {
+  public Entity mergeEntities(Entity primary, Entity secondary)
+      throws EuropeanaApiException, EntityModelCreationException {
 
     // TODO: consider refactoring of this implemeentation by creating a new class
     // EntityReconciliator
@@ -677,7 +691,7 @@ public class EntityRecordService {
    */
   private Entity combineEntities(
       Entity primary, Entity secondary, List<Field> fieldsToCombine, boolean accumulate)
-      throws EuropeanaApiException {
+      throws EuropeanaApiException, EntityModelCreationException {
     Entity consolidatedEntity =
         EntityObjectFactory.createConsolidatedEntityObject(primary.getType());
 
@@ -1043,16 +1057,14 @@ public class EntityRecordService {
   }
 
   /**
-   * @deprecated the service should add the deprecated filter, eventually by adding a boolean
-   *     parameter
+   * Fetches records matching the provided filter(s)
+   *
    * @param start
    * @param count
    * @param queryFilters
    * @return
    */
-  @Deprecated
-  public List<? extends EntityRecord> findEntitiesWithFilter(
-      int start, int count, Filter[] queryFilters) {
+  public List<EntityRecord> findEntitiesWithFilter(int start, int count, Filter[] queryFilters) {
     return this.entityRecordRepository.findWithFilters(start, count, queryFilters);
   }
 
@@ -1143,7 +1155,7 @@ public class EntityRecordService {
    * @throws EuropeanaApiException on exception
    */
   public void changeExternalProxy(EntityRecord entityRecord, String newProxyId)
-      throws EuropeanaApiException {
+      throws EuropeanaApiException, EntityModelCreationException {
 
     DataSource dataSource = datasources.verifyDataSource(newProxyId, true);
     List<EntityProxy> externalProxies = entityRecord.getExternalProxies();
